Register global variables
Global variable containing Ruby objects must be
declared with `rb_global_variable` so that the GC can
update the reference when it triggers compaction.

If it's not done it may cause the VM to crash.

In this instance it wasn't really a problem because
all the referenced objects are pinned, but it's still
good practice.